requirements.txt
matplotlib
numpy
skimage
django-rest-framework